Use bundle.js, DRY, create utils
