+++ public/index.html
@@ -6,7 +6,6 @@
     <meta name="description" content="A social and educational VR environment.">
     <script src="https://aframe.io/releases/0.4.0/aframe.min.js"></script>
     <script src="socket.io/socket.io.js"></script>
-    <script src="/components/publish-location.js"></script>
   </head>
   <body>
     <a-scene>
@@ -17,6 +16,6 @@
       <a-sky color="#ECECEC"></a-sky>
     </a-scene>
 
-    <script src="/index.js"></script>
+    <script src="/bundle.js"></script>
   </body>
 </html>
